param calculation_mode changed in risk config
The acceptable values of the parameter calculation_mode in the risk
configuration files have changed:
In OpenQuake v1.0, the values were classical, event_based and scenario;
In OpenQuake v1.2 the values are classical_risk, event_based_risk and
scenario_risk instead.
The names classical, event_based and scenario are now reserved for the
hazard calculators.