Fix missing author names in hazard bibliography
This is really a workaround. If two or more bibliography entries
have exactly the same authors in the same order, BibTeX replaces
the names of the authors with hyphens starting from the second
entry onwards. The workaround is to ensure that the author names
are not duplicated exactly in multiple entries. In this commit,
this is done by using full names in one entry, using initials in
another and so on.